add field validation and move nasdaq code into a folder
